+++ hedging/src/lib.rs
@@ -11,8 +11,7 @@ mod rebalance_action;
 mod synth_usd_liability;
 
 use galoy_client::GaloyClientConfig;
-use okex_client::OkexClientConfig;
-use shared::{health::HealthCheckTrigger, pubsub::*};
+use shared::{exchanges_config::OkexConfig, health::HealthCheckTrigger, payload::*, pubsub::*};
 
 pub use app::*;
 pub use error::*;
@@ -20,9 +19,10 @@ pub use error::*;
 pub async fn run(
     health_check_trigger: HealthCheckTrigger,
     config: HedgingAppConfig,
-    okex_config: OkexClientConfig,
+    okex_config: OkexConfig,
     galoy_config: GaloyClientConfig,
     pubsub_cfg: PubSubConfig,
+    tick_receiver: memory::Subscriber<PriceStreamPayload>,
 ) -> Result<(), HedgingError> {
     HedgingApp::run(
         health_check_trigger,
@@ -30,6 +30,7 @@ pub async fn run(
         okex_config,
         galoy_config,
         pubsub_cfg,
+        tick_receiver,
     )
     .await?;
     Ok(())