+++ okex-client/tests/client.rs
@@ -1,6 +1,7 @@
 use rust_decimal::Decimal;
 use rust_decimal_macros::dec;
 use serial_test::serial;
+use shared::exchanges_config::OkexConfig;
 
 use std::env;
 
@@ -11,7 +12,7 @@ async fn configured_okex_client() -> anyhow::Result<OkexClient> {
     let passphrase = env::var("OKEX_PASSPHRASE").expect("OKEX_PASS_PHRASE not set");
     let secret_key = env::var("OKEX_SECRET_KEY").expect("OKEX_SECRET_KEY not set");
 
-    let client = OkexClient::new(OkexClientConfig {
+    let client = OkexClient::new(OkexConfig {
         api_key,
         passphrase,
         secret_key,
@@ -50,7 +51,7 @@ async fn get_onchain_fees_data() -> anyhow::Result<()> {
 #[tokio::test]
 #[serial]
 async fn client_is_missing_header() -> anyhow::Result<()> {
-    let client = OkexClient::new(OkexClientConfig {
+    let client = OkexClient::new(OkexConfig {
         api_key: "".to_string(),
         passphrase: "".to_string(),
         secret_key: "".to_string(),