+++ okex-price/tests/price_feed.rs
@@ -1,27 +1,12 @@
 use chrono::Duration;
 use futures::StreamExt;
 use okex_price::*;
-use std::fs;
-use url::Url;
 
 use shared::{payload::*, pubsub::*, time::*};
 
-#[derive(serde::Deserialize)]
-struct Fixture {
-    payloads: Vec<PriceMessagePayload>,
-}
-
-fn load_fixture() -> anyhow::Result<Fixture> {
-    let contents =
-        fs::read_to_string("./tests/fixtures/price_feed.json").expect("Couldn't load fixtures");
-    Ok(serde_json::from_str(&contents)?)
-}
-
 #[tokio::test]
 async fn subscribes_to_tickers_channel() -> anyhow::Result<()> {
-    let config = PriceFeedConfig {
-        url: Url::parse("wss://ws.okx.com:8443/ws/v5/public").unwrap(),
-    };
+    let config = PriceFeedConfig::default();
     let mut received = subscribe_btc_usd_swap_price_tick(config)
         .await
         .expect("subscribe_btc_usd_swap");
@@ -47,9 +32,7 @@ async fn subscribes_to_tickers_channel() -> anyhow::Result<()> {
 
 #[tokio::test]
 async fn subscribe_to_order_book_channel() -> anyhow::Result<()> {
-    let config = PriceFeedConfig {
-        url: Url::parse("wss://ws.okx.com:8443/ws/v5/public").unwrap(),
-    };
+    let config = PriceFeedConfig::default();
     let mut order_book_stream = subscribe_btc_usd_swap_order_book(config)
         .await
         .expect("subscribe to order book channel");
@@ -70,31 +53,20 @@ async fn subscribe_to_order_book_channel() -> anyhow::Result<()> {
 }
 
 #[tokio::test]
-async fn publishes_to_redis() -> anyhow::Result<()> {
-    let redis_host = std::env::var("REDIS_HOST").unwrap_or("localhost".to_string());
-    let pubsub_config = PubSubConfig {
-        host: Some(redis_host),
-        ..PubSubConfig::default()
-    };
-    let mut subscriber = Subscriber::new(pubsub_config.clone()).await?;
+async fn publishes_to_price_stream() -> anyhow::Result<()> {
+    let (tick_send, mut tick_recv) =
+        memory::channel(chrono::Duration::from_std(std::time::Duration::from_secs(2)).unwrap());
 
     let _ = tokio::spawn(async move {
-        let _res = okex_price::run(PriceFeedConfig::default(), pubsub_config).await;
+        let _res = okex_price::run(PriceFeedConfig::default(), tick_send).await;
     });
 
-    let mut tick_stream = subscriber.subscribe::<OkexBtcUsdSwapPricePayload>().await?;
-    let mut book_stream = subscriber
-        .subscribe::<OkexBtcUsdSwapOrderBookPayload>()
-        .await?;
+    let received_tick = tick_recv.next().await.expect("expected price tick");
 
-    let received_tick = tick_stream.next().await.expect("expected price tick");
-    let received_book = book_stream.next().await.expect("expected price snapshot");
-
-    let payload = &load_fixture()?.payloads[0];
-    assert_eq!(received_tick.payload.exchange, payload.exchange);
-    assert_eq!(received_tick.payload.instrument_id, payload.instrument_id);
-    assert_eq!(received_book.payload.exchange, payload.exchange);
-    assert!(received_book.payload.asks.len() >= 400);
+    assert!(matches!(
+        received_tick.payload,
+        PriceStreamPayload::OkexBtcSwapPricePayload(_)
+    ));
 
     Ok(())
 }