+++ price-server/src/error.rs
@@ -1,7 +1,7 @@
 use thiserror::Error;
 
-use crate::{currency::CurrencyError, exchange_price_cache::ExchangePriceCacheError};
-use shared::pubsub::SubscriberError;
+use crate::currency::CurrencyError;
+use shared::{pubsub::SubscriberError, time::*};
 
 #[allow(clippy::large_enum_variant)]
 #[derive(Error, Debug)]
@@ -12,6 +12,14 @@ pub enum PriceAppError {
     SubscriberError(#[from] SubscriberError),
     #[error("PriceAppError - ExchangePriceCacheError: {0}")]
     ExchangePriceCacheError(#[from] ExchangePriceCacheError),
-    #[error("PriceAppError - StdDurationConversionError: {0}")]
-    StdDurationConversionError(#[from] chrono::OutOfRangeError),
+    #[error("PriceAppError - DecimalConversion: {0}")]
+    DecimalConversion(#[from] rust_decimal::Error),
+}
+
+#[derive(Error, Debug)]
+pub enum ExchangePriceCacheError {
+    #[error("StalePrice: last update was at {0}")]
+    StalePrice(TimeStamp),
+    #[error("No price data available")]
+    NoPriceAvailable,
 }