+++ price-server/src/lib.rs
@@ -2,16 +2,21 @@
 #![cfg_attr(feature = "fail-on-warnings", deny(clippy::all))]
 
 pub mod app;
+mod cache_config;
 pub mod currency;
-mod exchange_price_cache;
+mod error;
+mod exchange_tick_cache;
 mod fee_calculator;
+mod price_mixer;
 mod server;
 
-use shared::{health::HealthCheckTrigger, pubsub::PubSubConfig};
-
 use app::PriceApp;
+use shared::{
+    exchanges_config::ExchangeConfigs, health::HealthCheckTrigger, payload::*, pubsub::memory,
+};
+
 pub use app::PriceServerHealthCheckConfig;
-pub use exchange_price_cache::{ExchangePriceCacheConfig, ExchangePriceCacheError};
+pub use cache_config::ExchangePriceCacheConfig;
 pub use fee_calculator::FeeCalculatorConfig;
 pub use server::*;
 
@@ -20,15 +25,17 @@ pub async fn run(
     health_check_cfg: PriceServerHealthCheckConfig,
     server_config: PriceServerConfig,
     fee_calc_cfg: FeeCalculatorConfig,
-    pubsub_cfg: PubSubConfig,
+    subscriber: memory::Subscriber<PriceStreamPayload>,
     price_cache_config: ExchangePriceCacheConfig,
+    exchanges_cfg: ExchangeConfigs,
 ) -> Result<(), PriceServerError> {
     let app = PriceApp::run(
         health_check_trigger,
         health_check_cfg,
         fee_calc_cfg,
-        pubsub_cfg,
+        subscriber,
         price_cache_config,
+        exchanges_cfg,
     )
     .await?;