+++ price-server/tests/price_app.rs
@@ -1,9 +1,13 @@
-use futures::stream::StreamExt;
 use rust_decimal_macros::dec;
-use std::{fs, time::Duration};
+use std::fs;
 
-use price_server::{app::*, *};
-use shared::{payload::*, pubsub::*, time::*};
+use price_server::{app::*, ExchangePriceCacheConfig};
+use shared::{
+    exchanges_config::{ExchangeConfig, ExchangeConfigs, OkexConfig},
+    payload::*,
+    pubsub::*,
+    time::*,
+};
 
 #[derive(serde::Deserialize)]
 struct Fixture {
@@ -18,16 +22,28 @@ fn load_fixture() -> anyhow::Result<Fixture> {
 
 #[tokio::test]
 async fn price_app() -> anyhow::Result<()> {
-    let redis_host = std::env::var("REDIS_HOST").unwrap_or("localhost".to_string());
-    let config = PubSubConfig {
-        host: Some(redis_host),
-        ..PubSubConfig::default()
-    };
-    let publisher = Publisher::new(config.clone()).await?;
-    let mut subscriber = Subscriber::new(config.clone()).await?;
-    let mut stream = subscriber.subscribe::<OkexBtcUsdSwapPricePayload>().await?;
+    let (tick_send, tick_recv) =
+        memory::channel(chrono::Duration::from_std(std::time::Duration::from_secs(2)).unwrap());
+    let publisher = tick_send.clone();
+    let mut subscriber = tick_recv.resubscribe();
 
     let (_, recv) = futures::channel::mpsc::unbounded();
+
+    let okex = ExchangeConfig {
+        weight: dec!(1.0),
+        config: OkexConfig {
+            api_key: "okex api".to_string(),
+            passphrase: "passphrase".to_string(),
+            secret_key: "secret key".to_string(),
+            simulated: false,
+        },
+    };
+
+    let ex_cfgs = ExchangeConfigs {
+        okex: Some(okex),
+        kollider: None,
+    };
+
     let app = PriceApp::run(
         recv,
         PriceServerHealthCheckConfig::default(),
@@ -36,10 +52,9 @@ async fn price_app() -> anyhow::Result<()> {
             immediate_fee_rate: dec!(0.01),
             delayed_fee_rate: dec!(0.1),
         },
-        config,
-        ExchangePriceCacheConfig {
-            stale_after: Duration::from_secs(30),
-        },
+        tick_recv,
+        ExchangePriceCacheConfig::default(),
+        ex_cfgs,
     )
     .await?;
 
@@ -56,10 +71,10 @@ async fn price_app() -> anyhow::Result<()> {
 
     let mut payloads = load_fixture()?.payloads.into_iter();
     let mut payload = payloads.next().unwrap();
-    publisher
-        .publish(OkexBtcUsdSwapPricePayload(payload.clone()))
+    tick_send
+        .publish(PriceStreamPayload::OkexBtcSwapPricePayload(payload.clone()))
         .await?;
-    stream.next().await;
+    subscriber.next().await;
     tokio::time::sleep(std::time::Duration::from_millis(100)).await;
 
     let err = app
@@ -74,9 +89,9 @@ async fn price_app() -> anyhow::Result<()> {
 
     payload.timestamp = TimeStamp::now();
     publisher
-        .publish(OkexBtcUsdSwapPricePayload(payload))
+        .publish(PriceStreamPayload::OkexBtcSwapPricePayload(payload))
         .await?;
-    stream.next().await;
+    subscriber.next().await;
     tokio::time::sleep(std::time::Duration::from_millis(100)).await;
 
     let cents = app