+++ shared/src/payload/mod.rs
@@ -21,38 +21,13 @@ pub struct PriceMessagePayload {
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(transparent)]
-pub struct OkexBtcUsdSwapPricePayload(pub PriceMessagePayload);
-impl From<OkexBtcUsdSwapPricePayload> for PriceMessagePayload {
-    fn from(payload: OkexBtcUsdSwapPricePayload) -> Self {
-        payload.0
-    }
-}
-impl std::ops::Deref for OkexBtcUsdSwapPricePayload {
-    type Target = PriceMessagePayload;
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
+#[serde(tag = "type")]
+pub enum PriceStreamPayload {
+    OkexBtcSwapPricePayload(PriceMessagePayload),
+    KolliderBtcUsdSwapPricePayload(PriceMessagePayload),
 }
 
-crate::payload! { OkexBtcUsdSwapPricePayload, "price.okex.btc-usd-swap" }
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(transparent)]
-pub struct KolliderBtcUsdSwapPricePayload(pub PriceMessagePayload);
-impl From<KolliderBtcUsdSwapPricePayload> for PriceMessagePayload {
-    fn from(payload: KolliderBtcUsdSwapPricePayload) -> Self {
-        payload.0
-    }
-}
-impl std::ops::Deref for KolliderBtcUsdSwapPricePayload {
-    type Target = PriceMessagePayload;
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-crate::payload! { KolliderBtcUsdSwapPricePayload, "price.kollider.btc-usd-swap" }
+crate::payload! { PriceStreamPayload, "price.stream" }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct SynthUsdLiabilityPayload {